Split process-rich function
The process-rich function was developed for entries in the 'outputs'
field of a cell. This is a list of objects with each a 'data' field,
alongside a 'metadata' and 'output_type' field.

The new process-rich function only processes the 'data' field, which is
a general dict of all possible representations.
The point of splitting this function is to re-use this smaller-scoped
function when processing 'attachments'.

The functionality of the original process-rich is moved to the
process-rich-output function.