Pass through extra args to handlers
When extra arguments are passed to read-mime, pass them to the handler.
This prepares for when extra arguments are passed to e.g. the markdown
handler. If extra arguments are passed to a handler, then user-defined
handlers will break (i.e., when this mechanism is used in the future,
it will be a breaking change).