Add image-markdown-cell to handle images in markdown
This function handles embedded images ('attachments') in markdown, by
reusing the existing process-rich function and handlers. This uses
mutual recursion, which is solved by passing the required functions in
the handlers dict.

If the image is not embedded, the path is delegated to the standard
image function. Since Typst doesn't have a 'path' type yet, it is a
necessity that this is also done by a handler at the key "image/x.path".
To function correctly, the user needs to overwrite this key (a next
commit will update the documentation).

Already adds the corresponding handlers.
The function is also already used in the cmarker scope, but this doesn't
change anything since the 'attachment' path is not yet triggered.